requirements.txt
APScheduler==3.6.3
autopep8==2.0.0
cachetools==4.2.2
certifi==2022.9.24
greenlet==1.1.3.post0
pycodestyle==2.9.1
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.42
tomli==2.0.1
tornado==6.1
tzdata==2022.4
tzlocal==4.2
